Emit modification when a field is added/removed
We need to treat addition and removal as breaking change because it may
affect match patterns.

Previous implementation treated struct and enum fields as a separate
requirement, completely detached from their type. This caused that a
field addition was reported as a non-breaking change, which is not
correct.

This commit fixes the problem by counting the fields as a metadata of
their datatype. By doing so, every new field, type modification or field
removal is reported as a breaking change.